requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
requests==2.31.0
beautifulsoup4==4.12.2
PyMySQL==1.1.0
python-dotenv==1.0.0
openai==1.3.3
anthropic==0.3.11
google-generativeai==0.3.0
pandas==2.1.1
lxml==4.9.3
cssselect==1.2.0